python/django/backend/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
build==1.2.1
CacheControl==0.14.0
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crashtest==0.4.1
cron-descriptor==1.4.3
distlib==0.3.8
Django==4.2.11
django-celery-beat==2.6.0
django-ckeditor-5==0.2.13
django-cleanup==8.1.0
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-js-asset==2.2.0
django-mptt==0.16.0
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-simple-captcha==0.6.0
django-taggit==5.0.1
django-timezone-field==6.1.0
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.14.0
idna==3.7
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
kombu==5.3.7
more-itertools==10.2.0
msgpack==1.0.8
packaging==24.0
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.11.0
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pyproject_hooks==1.1.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytils==0.4.1
rapidfuzz==3.9.3
redis==5.0.7
requests==2.32.3
requests-toolbelt==1.0.0
shellingham==1.5.4
six==1.16.0
sqlparse==0.5.0
tomlkit==0.12.5
trove-classifiers==2024.5.22
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.26.2
wcwidth==0.2.13
xattr==1.1.0
gunicorn